+++ rpc/apis.go
@@ -18,11 +18,11 @@ package rpc
 import (
 	"fmt"
 
+	rpcclient "github.com/cometbft/cometbft/rpc/jsonrpc/client"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/server"
 	"github.com/ethereum/go-ethereum/rpc"
 	ethermint "github.com/evmos/ethermint/types"
-	rpcclient "github.com/tendermint/tendermint/rpc/jsonrpc/client"
 	"github.com/zeta-chain/zetacore/rpc/backend"
 	"github.com/zeta-chain/zetacore/rpc/namespaces/ethereum/debug"
 	"github.com/zeta-chain/zetacore/rpc/namespaces/ethereum/eth"