+++ testutil/keeper/keeper.go
@@ -5,6 +5,10 @@ import (
 	"testing"
 	"time"
 
+	tmdb "github.com/cometbft/cometbft-db"
+	"github.com/cometbft/cometbft/crypto/tmhash"
+	"github.com/cometbft/cometbft/libs/log"
+	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/store"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
@@ -13,6 +17,8 @@ import (
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	consensuskeeper "github.com/cosmos/cosmos-sdk/x/consensus/keeper"
+	consensustypes "github.com/cosmos/cosmos-sdk/x/consensus/types"
 	distrkeeper "github.com/cosmos/cosmos-sdk/x/distribution/keeper"
 	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
@@ -32,10 +38,6 @@ import (
 	feemarketkeeper "github.com/evmos/ethermint/x/feemarket/keeper"
 	feemarkettypes "github.com/evmos/ethermint/x/feemarket/types"
 	"github.com/stretchr/testify/require"
-	"github.com/tendermint/tendermint/crypto/tmhash"
-	"github.com/tendermint/tendermint/libs/log"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
-	tmdb "github.com/tendermint/tm-db"
 	"github.com/zeta-chain/zetacore/testutil/sample"
 	authoritymodule "github.com/zeta-chain/zetacore/x/authority"
 	authoritykeeper "github.com/zeta-chain/zetacore/x/authority/keeper"
@@ -146,12 +148,22 @@ func ParamsKeeper(
 	)
 }
 
+func ConsensusKeeper(
+	cdc codec.Codec,
+	db *tmdb.MemDB,
+	ss store.CommitMultiStore,
+) consensuskeeper.Keeper {
+	storeKey := sdk.NewKVStoreKey(consensustypes.StoreKey)
+
+	ss.MountStoreWithDB(storeKey, storetypes.StoreTypeIAVL, db)
+	return consensuskeeper.NewKeeper(cdc, storeKey, authtypes.NewModuleAddress(govtypes.ModuleName).String())
+}
+
 // AccountKeeper instantiates an account keeper for testing purposes
 func AccountKeeper(
 	cdc codec.Codec,
 	db *tmdb.MemDB,
 	ss store.CommitMultiStore,
-	paramKeeper paramskeeper.Keeper,
 ) authkeeper.AccountKeeper {
 	storeKey := sdk.NewKVStoreKey(authtypes.StoreKey)
 	ss.MountStoreWithDB(storeKey, storetypes.StoreTypeIAVL, db)
@@ -159,10 +171,10 @@ func AccountKeeper(
 	return authkeeper.NewAccountKeeper(
 		cdc,
 		storeKey,
-		paramKeeper.Subspace(authtypes.ModuleName),
 		ethermint.ProtoAccount,
 		moduleAccountPerms,
 		"zeta",
+		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
 	)
 }
 
@@ -171,7 +183,6 @@ func BankKeeper(
 	cdc codec.Codec,
 	db *tmdb.MemDB,
 	ss store.CommitMultiStore,
-	paramKeeper paramskeeper.Keeper,
 	authKeeper authkeeper.AccountKeeper,
 ) bankkeeper.Keeper {
 	storeKey := sdk.NewKVStoreKey(banktypes.StoreKey)
@@ -182,8 +193,8 @@ func BankKeeper(
 		cdc,
 		storeKey,
 		authKeeper,
-		paramKeeper.Subspace(banktypes.ModuleName),
 		blockedAddrs,
+		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
 	)
 }
 
@@ -194,25 +205,24 @@ func StakingKeeper(
 	ss store.CommitMultiStore,
 	authKeeper authkeeper.AccountKeeper,
 	bankKeeper bankkeeper.Keeper,
-	paramKeeper paramskeeper.Keeper,
 ) stakingkeeper.Keeper {
 	storeKey := sdk.NewKVStoreKey(stakingtypes.StoreKey)
 	ss.MountStoreWithDB(storeKey, storetypes.StoreTypeIAVL, db)
 
-	return stakingkeeper.NewKeeper(
+	return *stakingkeeper.NewKeeper(
 		cdc,
 		storeKey,
 		authKeeper,
 		bankKeeper,
-		paramKeeper.Subspace(stakingtypes.ModuleName),
+		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
 	)
 }
 
 // SlashingKeeper instantiates a slashing keeper for testing purposes
-func SlashingKeeper(cdc codec.Codec, db *tmdb.MemDB, ss store.CommitMultiStore, stakingKeeper stakingkeeper.Keeper, paramKeeper paramskeeper.Keeper) slashingkeeper.Keeper {
+func SlashingKeeper(cdc codec.Codec, db *tmdb.MemDB, ss store.CommitMultiStore, stakingKeeper stakingkeeper.Keeper) slashingkeeper.Keeper {
 	storeKey := sdk.NewKVStoreKey(slashingtypes.StoreKey)
 	ss.MountStoreWithDB(storeKey, storetypes.StoreTypeIAVL, db)
-	return slashingkeeper.NewKeeper(cdc, storeKey, stakingKeeper, paramKeeper.Subspace(slashingtypes.ModuleName))
+	return slashingkeeper.NewKeeper(cdc, codec.NewLegacyAmino(), storeKey, stakingKeeper, authtypes.NewModuleAddress(govtypes.ModuleName).String())
 }
 
 // DistributionKeeper instantiates a distribution keeper for testing purposes
@@ -223,7 +233,6 @@ func DistributionKeeper(
 	authKeeper authkeeper.AccountKeeper,
 	bankKeeper bankkeeper.Keeper,
 	stakingKeeper *stakingkeeper.Keeper,
-	paramKeeper paramskeeper.Keeper,
 ) distrkeeper.Keeper {
 	storeKey := sdk.NewKVStoreKey(distrtypes.StoreKey)
 	ss.MountStoreWithDB(storeKey, storetypes.StoreTypeIAVL, db)
@@ -231,11 +240,11 @@ func DistributionKeeper(
 	return distrkeeper.NewKeeper(
 		cdc,
 		storeKey,
-		paramKeeper.Subspace(stakingtypes.ModuleName),
 		authKeeper,
 		bankKeeper,
 		stakingKeeper,
 		authtypes.FeeCollectorName,
+		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
 	)
 }
 
@@ -256,7 +265,7 @@ func UpgradeKeeper(
 	skipUpgradeHeights := make(map[int64]bool)
 	vs := ProtocolVersionSetter{}
 
-	return upgradekeeper.NewKeeper(
+	return *upgradekeeper.NewKeeper(
 		skipUpgradeHeights,
 		storeKey,
 		cdc,
@@ -272,6 +281,7 @@ func FeeMarketKeeper(
 	db *tmdb.MemDB,
 	ss store.CommitMultiStore,
 	paramKeeper paramskeeper.Keeper,
+	consensusKeeper consensuskeeper.Keeper,
 ) feemarketkeeper.Keeper {
 	storeKey := sdk.NewKVStoreKey(feemarkettypes.StoreKey)
 	transientKey := sdk.NewTransientStoreKey(feemarkettypes.TransientKey)
@@ -285,6 +295,7 @@ func FeeMarketKeeper(
 		storeKey,
 		transientKey,
 		paramKeeper.Subspace(feemarkettypes.ModuleName),
+		consensusKeeper,
 	)
 }
 
@@ -298,6 +309,7 @@ func EVMKeeper(
 	stakingKeeper stakingkeeper.Keeper,
 	feemarketKeeper feemarketkeeper.Keeper,
 	paramKeeper paramskeeper.Keeper,
+	consensusKeeper consensuskeeper.Keeper,
 ) *evmkeeper.Keeper {
 	storeKey := sdk.NewKVStoreKey(evmtypes.StoreKey)
 	transientKey := sdk.NewTransientStoreKey(evmtypes.TransientKey)
@@ -318,6 +330,7 @@ func EVMKeeper(
 		geth.NewEVM,
 		"",
 		paramKeeper.Subspace(evmtypes.ModuleName),
+		consensusKeeper,
 	)
 
 	return k
@@ -330,12 +343,13 @@ func NewSDKKeepers(
 	ss store.CommitMultiStore,
 ) SDKKeepers {
 	paramsKeeper := ParamsKeeper(cdc, db, ss)
-	authKeeper := AccountKeeper(cdc, db, ss, paramsKeeper)
-	bankKeeper := BankKeeper(cdc, db, ss, paramsKeeper, authKeeper)
-	stakingKeeper := StakingKeeper(cdc, db, ss, authKeeper, bankKeeper, paramsKeeper)
-	feeMarketKeeper := FeeMarketKeeper(cdc, db, ss, paramsKeeper)
-	evmKeeper := EVMKeeper(cdc, db, ss, authKeeper, bankKeeper, stakingKeeper, feeMarketKeeper, paramsKeeper)
-	slashingKeeper := SlashingKeeper(cdc, db, ss, stakingKeeper, paramsKeeper)
+	authKeeper := AccountKeeper(cdc, db, ss)
+	bankKeeper := BankKeeper(cdc, db, ss, authKeeper)
+	stakingKeeper := StakingKeeper(cdc, db, ss, authKeeper, bankKeeper)
+	consensusKeeper := ConsensusKeeper(cdc, db, ss)
+	feeMarketKeeper := FeeMarketKeeper(cdc, db, ss, paramsKeeper, consensusKeeper)
+	evmKeeper := EVMKeeper(cdc, db, ss, authKeeper, bankKeeper, stakingKeeper, feeMarketKeeper, paramsKeeper, consensusKeeper)
+	slashingKeeper := SlashingKeeper(cdc, db, ss, stakingKeeper)
 	return SDKKeepers{
 		ParamsKeeper:    paramsKeeper,
 		AuthKeeper:      authKeeper,